Создание модулей utils.js и const.js. Вынесем повторяющиеся данные в модуль с константами. Вынесем вспомогательные функции и общие функции в модуль с утилитами.
